+++ src/systems/block_attach/magnet.rs
@@ -46,36 +46,44 @@ pub fn spawn_magnet(commands: &mut Commands, block_ent: Entity, rect: &Rectangle
             ..Default::default()
         },
     );
-    commands.entity(block_ent).with_children(|parent| {
-        parent.spawn_bundle(magnet_shape_bundle);
-    });
+    let child_ent = commands
+        .spawn_bundle(magnet_shape_bundle)
+        .insert(magnet.clone())
+        .id();
+    commands.entity(block_ent).push_children(&[child_ent]);
 }
 
 /// 磁石とボールの間に力を加える
 fn magnet_force(
-    magnet_query: Query<(&Magnet, &BlockTransform, &GlobalTransform, &BlockType)>,
+    block_query: Query<(&BlockTransform, &GlobalTransform, &BlockType, &Children)>,
+    magnet_query: Query<&Magnet>,
     mut ball_query: Query<(&Ball, &Position, &mut Force), With<MetalBall>>,
 ) {
-    for (magnet, block_trans, block_glb_trans, block_type) in magnet_query.iter() {
-        if magnet.active {
-            if let BlockType::Rect { shape } = block_type {
-                let angle = block_trans.angle;
-                let (_, _, block_glb_translation) = block_glb_trans.to_scale_rotation_translation();
-                let [p1, p2] = calc_edge_points_of_rectangle(
-                    &magnet.direction,
-                    block_glb_translation.truncate(),
-                    angle,
-                    shape.extents,
-                );
+    for (block_trans, block_glb_trans, block_type, block_children) in block_query.iter() {
+        for &child in block_children.iter() {
+            if let Ok(magnet) = magnet_query.get(child) {
+                if magnet.active {
+                    if let BlockType::Rect { shape } = block_type {
+                        let angle = block_trans.angle;
+                        let (_, _, block_glb_translation) =
+                            block_glb_trans.to_scale_rotation_translation();
+                        let [p1, p2] = calc_edge_points_of_rectangle(
+                            &magnet.direction,
+                            block_glb_translation.truncate(),
+                            angle,
+                            shape.extents,
+                        );
 
-                for (_, ball_pos, mut force) in ball_query.iter_mut() {
-                    let ball_pos = ball_pos.0;
-                    if (p2 - p1).dot(ball_pos - p1) > 0.0
-                        && (p1 - p2).dot(ball_pos - p2) > 0.0
-                        && (ball_pos - p1).perp_dot(p2 - p1) > 0.0
-                    {
-                        let dir_unit = (p2 - p1).perp().normalize();
-                        force.0 += dir_unit * magnet.flux_density * magnet.flux_density;
+                        for (_, ball_pos, mut force) in ball_query.iter_mut() {
+                            let ball_pos = ball_pos.0;
+                            if (p2 - p1).dot(ball_pos - p1) > 0.0
+                                && (p1 - p2).dot(ball_pos - p2) > 0.0
+                                && (ball_pos - p1).perp_dot(p2 - p1) > 0.0
+                            {
+                                let dir_unit = (p2 - p1).perp().normalize();
+                                force.0 += dir_unit * magnet.flux_density * magnet.flux_density;
+                            }
+                        }
                     }
                 }
             }
